Add JobAdvertisementController Impl

+++ waa-backend/src/main/java/edu/miu/waabackend/service/IJobAdvertisementService.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface IJobAdvertisementService {
     public List<DTOEntity> lstGetAll();
-    public DTOEntity GetByPK(Long id);
-    public int Insert(DTOEntity jobAdvertisementDto);
-    public int Update(DTOEntity jobAdvertisementDto);
-    public int Delete(Long id);
+    public DTOEntity GetByPK(long id);
+    public long Insert(DTOEntity jobAdvertisementDto);
+    public long Update(DTOEntity jobAdvertisementDto);
+    public long Delete(long id);
 }
